fix: 채팅 리스트 loading 개선
